+++ processing-api/src/main/java/org/n52/javaps/algorithm/annotation/LiteralInput.java
@@ -38,8 +38,8 @@ public @interface LiteralInput {
     String identifier(); // identifier
     String title() default "";
     String abstrakt() default ""; // 'abstract' is java reserved keyword
-    int minOccurs() default 1;
-    int maxOccurs() default 1;
+    long minOccurs() default 1;
+    long maxOccurs() default 1;
     String defaultValue() default "";
     String[] allowedValues() default {};
     String uom() default "";
